+++ lib/constants/index.js
@@ -11,7 +11,7 @@ const PHP_ENVIRONMENT = 'php'
 const GO_ENVIRONMENT = 'go'
 const JAVA_ENVIRONMENT = 'java'
 
-const TEMPLATE_PATH = '../../templates'
+const TEMPLATE_PATH = '../../../templates'
 
 const ASSIGN_WORKFLOW = 'Assign'
 const LINTER_WORKFLOW = 'Linter'
